added models for js mapping

+++ src/Conclave.Oracle.Node/Models/INetwork.cs
@@ -1,10 +1,12 @@
+using System.Numerics;
+
 namespace Conclave.Oracle.Node.Interfaces;
 
 public interface INetwork
 {
-    ulong ZeroTime { get; set; }
-    ulong ZeroSlot { get; set; }
-    ulong SlotLength { get; set; }
+    BigInteger ZeroTime { get; set; }
+    BigInteger ZeroSlot { get; set; }
+    BigInteger SlotLength { get; set; }
 
-    int UnixTimeMsToSlot(ulong unixTimeMs);
+    int UnixTimeMsToSlot(BigInteger unixTimeMs);
 }
